third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFETurbulenceElement-dom-baseFrequency-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'baseFrequency' attribute of the SVGFETurbulenceElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS turbulence.getAttribute('baseFrequency') is "0.07"
PASS turbulence.getAttribute('baseFrequency') is "0.05"
PASS successfullyParsed is true

TEST COMPLETE
